+++ tests/native/src/test/scala/cron4s/NativeCronSpec.scala
@@ -16,10 +16,11 @@
 
 package cron4s
 
-package object parsing {
-  private[cron4s] def parse(e: String): Either[Error, CronExpr] =
-    for {
-      tokens <- CronLexer.tokenize(e)
-      expr   <- CronParser.read(tokens)
-    } yield expr
+import org.scalatest.flatspec.AnyFlatSpec
+
+import scala.annotation.nowarn
+
+class ParserCombinatorsCronSpec extends AnyFlatSpec with CronSpec {
+  @nowarn("cat=deprecation")
+  def parser = parsing.Parser
 }